Rearrange markdown module
